refactor(frontend): split lib/api.ts into per-resource lib/api/ modules
Frontend components previously imported every API function from one
2500-line file. Split it into lib/api/client.ts (axios instance,
CSRF/JWT session handling, error-message mapping) plus one file per
backend resource (jobs, applications, profiles, escrow, messages,
etc.), all still funneling through the shared client. lib/api/index.ts
re-exports everything so existing `from "@/lib/api"` imports keep
working unchanged.

Closes #830
