--- frontend/lib/api.ts
@@ -1,2565 +1,0 @@
-import axios from "axios";
-import { optionalClientEnv } from "./env";
-import type {
-  ClientReputation,
-  Availability,
-  Job,
-  Application,
-  JobAnalytics,
-  UserProfile,
-  Rating,
-  ProposalTemplate,
-  PriceAlertPreference,
-  SkillEndorsement,
-  ClientSpendingAnalytics,
-  PortfolioFile,
-  TokenInfo,
-  TokenBalance,
-  TimeEntry,
-  TimeInvoice,
-  Message,
-  ReferralStats,
-  AssessmentQuestion,
-  SkillBadge,
-  BulkActionResponse,
-  NotificationItem,
-  ProfileStats,
-  ResponseTime,
-  AuditLogEntry,
-} from "@/utils/types";
-
-// ─── Structured error code → user-facing string map (#461) ───────────────────
-export const API_ERROR_MESSAGES: Record<string, string> = {
-  INTERNAL_SERVER_ERROR:      "Something went wrong. Please try again.",
-  VALIDATION_ERROR:           "Invalid input. Please check your data.",
-  NOT_FOUND:                  "The requested resource was not found.",
-  UNAUTHORIZED:               "You need to sign in to do that.",
-  FORBIDDEN:                  "You don't have permission to do that.",
-  RATE_LIMITED:               "Too many requests. Please slow down.",
-  BAD_REQUEST:                "Bad request. Please check your input.",
-  INVALID_TOKEN:              "Your session is invalid. Please sign in again.",
-  TOKEN_EXPIRED:              "Your session has expired. Please sign in again.",
-  ADDRESS_MISMATCH:           "Wallet address does not match.",
-  PROFILE_NOT_FOUND:          "Profile not found.",
-  PROFILE_DELETED:            "This profile has been deleted.",
-  ENCRYPTION_KEY_INVALID:     "Invalid encryption key format.",
-  JOB_NOT_FOUND:              "Job not found.",
-  JOB_ALREADY_EXPIRED:        "This job has already expired.",
-  JOB_NOT_OPEN:               "This job is no longer accepting applications.",
-  APPLICATION_NOT_FOUND:      "Application not found.",
-  ALREADY_APPLIED:            "You have already applied to this job.",
-  ESCROW_ALREADY_EXISTS:      "An escrow already exists for this job.",
-  ESCROW_NOT_FOUND:           "Escrow not found.",
-  INSUFFICIENT_BALANCE:       "Insufficient balance for this operation.",
-  ESCROW_ALREADY_RELEASED:    "Escrow has already been released.",
-  ESCROW_TIMEOUT_NOT_REACHED: "The escrow timeout period has not been reached yet.",
-  DISPUTE_NOT_FOUND:          "Dispute not found.",
-  DISPUTE_ALREADY_EXISTS:     "A dispute already exists for this job.",
-  EVIDENCE_LIMIT_REACHED:     "You have reached the maximum number of evidence files.",
-  EVIDENCE_NOT_FOUND:         "Evidence file not found.",
-  MESSAGE_NOT_FOUND:          "Message not found.",
-  MESSAGE_TOO_LONG:           "Message exceeds the maximum length.",
-  NOT_JOB_PARTICIPANT:        "You are not a participant in this job.",
-  FILE_TOO_LARGE:             "File is too large.",
-  FILE_TYPE_NOT_ALLOWED:      "This file type is not allowed.",
-  PORTFOLIO_LIMIT_REACHED:    "You have reached the maximum number of portfolio items.",
-  IPFS_UPLOAD_FAILED:         "File upload failed. Please try again.",
-  PINATA_NOT_CONFIGURED:      "File storage is temporarily unavailable.",
-  SIGNED_URL_EXPIRED:         "This download link has expired. Please request a new one.",
-  SIGNED_URL_INVALID:         "This download link is invalid.",
-  JSONB_DEPTH_EXCEEDED:       "Input data is too deeply nested.",
-  JSONB_SCHEMA_INVALID:       "Input data does not match the expected format.",
-};
-
-/**
- * Extract a user-facing message from an API error response.
- * Handles both the new structured { error: { code, message } } shape
- * and the legacy { error: "string" } shape.
- */
-export function getApiErrorMessage(error: unknown, fallback = "An unexpected error occurred."): string {
-  if (!axios.isAxiosError(error)) return fallback;
-  const data = error.response?.data;
-  if (!data) return fallback;
-  // Structured shape
-  if (data.error && typeof data.error === "object" && data.error.code) {
-    return API_ERROR_MESSAGES[data.error.code] ?? data.error.message ?? fallback;
-  }
-  // Legacy shape
-  if (typeof data.error === "string") return data.error;
-  return fallback;
-}
-
-const api = axios.create({
-  baseURL: optionalClientEnv("NEXT_PUBLIC_API_URL", "http://localhost:4000"),
-  headers: { "Content-Type": "application/json" },
-  withCredentials: true,
-  timeout: 10000,
-});
-
-let jwtToken: string | null = null;
-let refreshTimer: ReturnType<typeof setTimeout> | null = null;
-let refreshPromise: Promise<string | null> | null = null;
-
-// ── Request tracing (Issue #453) ────────────────────────────────────────────
-const REQUEST_ID_HEADER = "X-Request-ID";
-const IS_DEV = process.env.NODE_ENV !== "production";
-
-/** Generate a UUID v4 using crypto.randomUUID when available, with RFC4122 fallback. */
-function generateRequestId(): string {
-  try {
-    if (typeof crypto !== "undefined" && typeof crypto.randomUUID === "function") {
-      return crypto.randomUUID();
-    }
-  } catch {
-    // fall through to manual generator
-  }
-  // Fallback for environments without crypto.randomUUID (older browsers / SSR):
-  const bytes = new Uint8Array(16);
-  if (typeof crypto !== "undefined" && typeof crypto.getRandomValues === "function") {
-    crypto.getRandomValues(bytes);
-  } else {
-    for (let i = 0; i < bytes.length; i += 1) bytes[i] = Math.floor(Math.random() * 256);
-  }
-  // Per RFC 4122 §4.4
-  bytes[6] = (bytes[6] & 0x0f) | 0x40;
-  bytes[8] = (bytes[8] & 0x3f) | 0x80;
-  const hex: string[] = [];
-  for (let i = 0; i < bytes.length; i += 1) {
-    hex.push(bytes[i].toString(16).padStart(2, "0"));
-  }
-  return (
-    `${hex.slice(0, 4).join("")}-${hex.slice(4, 6).join("")}-${hex[6]}-${hex[7]}-` +
-    `${hex.slice(8, 10).join("")}-${hex.slice(10, 16).join("")}`
-  );
-}
-
-/**
- * Capture the server-issued X-Request-ID from a successful response so
- * components can surface it in error toasts / support tickets. Returns
- * `null` when the header is absent (older backends).
- */
-export function getResponseRequestId(error: unknown): string | null {
-  if (!axios.isAxiosError(error)) return null;
-  return error.response?.headers?.[REQUEST_ID_HEADER.toLowerCase()] ?? null;
-}
-
-// ── CSRF (Issue #451) ────────────────────────────────────────────────────────
-const CSRF_HEADER = "X-CSRF-Token";
-const CSRF_TOKEN_URL = "/api/auth/csrf-token";
-const MUTATING_METHODS = new Set(["post", "put", "patch", "delete"]);
-
-let csrfToken: string | null = null;
-let csrfFetchPromise: Promise<string | null> | null = null;
-
-async function fetchCsrfToken(): Promise<string | null> {
-  if (csrfFetchPromise) return csrfFetchPromise;
-  // /api/auth/csrf-token is exempt from CSRF (it's a safe-method bootstrap
-  // endpoint), so we do NOT set the X-CSRF-Token header on this request —
-  // doing so would risk an infinite loop if the server refused it.
-  csrfFetchPromise = api
-    .get<{ csrfToken: string }>(CSRF_TOKEN_URL, { skipCsrf: true } as any)
-    .then(({ data }) => {
-      csrfToken = data.csrfToken;
-      return csrfToken;
-    })
-    .catch(() => {
-      csrfToken = null;
-      return null;
-    })
-    .finally(() => {
-      csrfFetchPromise = null;
-    });
-  return csrfFetchPromise;
-}
-
-export function clearCsrfToken() {
-  csrfToken = null;
-}
-
-/**
- * Invalidate the cached token and mint a fresh one. Used by the response
- * interceptor when a mutation rejects with 403 (cookie mismatch) and by
- * `verifyAuthChallenge` / `refreshAccessToken` when the server returns a
- * new token in-band. Deduplicates concurrent callers to avoid racing
- * parallel mutations against different cookie versions.
- */
-async function refreshCsrfToken(): Promise<string | null> {
-  clearCsrfToken();
-  return fetchCsrfToken();
-}
-
-const TOKEN_REFRESH_BUFFER_MS = 60 * 1000;
-
-function getJwtExpiryMs(token: string) {
-  try {
-    const encodedPayload = token.split(".")[1] || "";
-    const base64Payload = encodedPayload
-      .replace(/-/g, "+")
-      .replace(/_/g, "/")
-      .padEnd(Math.ceil(encodedPayload.length / 4) * 4, "=");
-    const payload = JSON.parse(atob(base64Payload));
-    return typeof payload.exp === "number" ? payload.exp * 1000 : null;
-  } catch {
-    return null;
-  }
-}
-
-function clearRefreshTimer() {
-  if (refreshTimer) {
-    clearTimeout(refreshTimer);
-    refreshTimer = null;
-  }
-}
-
-function scheduleTokenRefresh(token: string) {
-  if (typeof window === "undefined") return;
-
-  const expiryMs = getJwtExpiryMs(token);
-  if (!expiryMs) return;
-
-  const delayMs = Math.max(expiryMs - Date.now() - TOKEN_REFRESH_BUFFER_MS, 0);
-  refreshTimer = setTimeout(() => {
-    refreshAccessToken().catch(() => setJwtToken(null));
-  }, delayMs);
-}
-
-function shouldRefreshToken() {
-  if (!jwtToken) return false;
-  const expiryMs = getJwtExpiryMs(jwtToken);
-  return Boolean(expiryMs && expiryMs - Date.now() <= TOKEN_REFRESH_BUFFER_MS);
-}
-
-export function setJwtToken(token: string | null) {
-  jwtToken = token;
-  clearRefreshTimer();
-  if (token) scheduleTokenRefresh(token);
-}
-
-export function getJwtToken() {
-  return jwtToken;
-}
-
-api.interceptors.request.use(async (config: any) => {
-  if (!config.skipAuthRefresh && shouldRefreshToken()) {
-    await refreshAccessToken();
-  }
-
-  // Issue #453: attach a request id so server logs can be correlated with
-  // client-side debugging. Generated fresh per request unless caller
-  // supplied one via `config.requestId` (useful for tracing flows).
-  if (!config.skipTracing && !config.headers?.[REQUEST_ID_HEADER]) {
-    config.headers = config.headers || {};
-    config.headers[REQUEST_ID_HEADER] = config.requestId || generateRequestId();
-  }
-
-  // Attach the CSRF token to every mutating request so the backend's
-  // double-submit cookie check passes (Issue #451).
-  if (!config.skipCsrf && MUTATING_METHODS.has((config.method || "").toLowerCase())) {
-    if (!csrfToken) await fetchCsrfToken();
-    if (csrfToken) {
-      config.headers = config.headers || {};
-      config.headers[CSRF_HEADER] = csrfToken;
-    }
-  }
-  return config;
-});
-
-api.interceptors.response.use(
-  (response) => {
-    // Issue #453 dev-mode logging: print the correlation id so a developer
-    // can paste it straight into backend log search.
-    if (IS_DEV && typeof console !== "undefined") {
-      const serverId =
-        response.headers?.[REQUEST_ID_HEADER.toLowerCase()] ||
-        response.config?.headers?.[REQUEST_ID_HEADER];
-      if (serverId) {
-        // eslint-disable-next-line no-console
-        console.debug(`[api] ${response.config?.method?.toUpperCase()} ${response.config?.url} → ${response.status} requestId=${serverId}`);
-      }
-    }
-    return response;
-  },
-  async (error) => {
-    // Capture the request id on errors too so error toasts can show it.
-    if (IS_DEV && typeof console !== "undefined" && axios.isAxiosError(error)) {
-      const serverId = error.response?.headers?.[REQUEST_ID_HEADER.toLowerCase()];
-      const clientId = error.config?.headers?.[REQUEST_ID_HEADER];
-      // eslint-disable-next-line no-console
-      console.warn(
-        `[api] ${error.config?.method?.toUpperCase()} ${error.config?.url} → ${error.response?.status || "NETWORK"} ` +
-          `requestId=${serverId || clientId || "?"}`,
-      );
-    }
-
-    const originalRequest = error.config || {};
-    // 401 → refresh the JWT once and replay (existing behavior).
-    if (
-      error.response?.status === 401 &&
-      !originalRequest._retry &&
-      !originalRequest.skipAuthRefresh
-    ) {
-      originalRequest._retry = true;
-      const token = await refreshAccessToken();
-      if (token) {
-        return api(originalRequest);
-      }
-    }
-    // 403 on a mutating request → CSRF token may have been rotated or the
-    // cookie was replaced by another tab. Mint a fresh pair and retry once.
-    // We route through refreshCsrfToken() so concurrent 403s share a single
-    // in-flight mint instead of all racing the server independently.
-    if (
-      error.response?.status === 403 &&
-      !originalRequest._csrfRetry &&
-      !originalRequest.skipCsrf &&
-      MUTATING_METHODS.has((originalRequest.method || "").toLowerCase())
-    ) {
-      originalRequest._csrfRetry = true;
-      const fresh = await refreshCsrfToken();
-      if (fresh) {
-        originalRequest.headers = originalRequest.headers || {};
-        originalRequest.headers[CSRF_HEADER] = fresh;
-        return api(originalRequest);
-      }
-    }
-    throw error;
-  },
-);
-
-// ─── Auth ─────────────────────────────────────────────────────────────────────
-
-export async function fetchAuthChallenge(publicKey: string) {
-  const { data } = await api.get<{ transaction: string }>(
-    `/api/auth?account=${publicKey}`,
-  );
-  return data.transaction;
-}
-
-export async function verifyAuthChallenge(transaction: string) {
-  const { data } = await api.post<{ success: boolean; token: string; csrfToken?: string }>(
-    "/api/auth",
-    { transaction },
-  );
-  // The login response may ship a fresh CSRF token alongside the JWT so we
-  // can skip the post-login /api/auth/csrf-token round-trip (#451).
-  if (data.csrfToken) csrfToken = data.csrfToken;
-  return data.token;
-}
-
-export async function refreshAccessToken() {
-  if (!refreshPromise) {
-    refreshPromise = api
-      .post<{ success: boolean; token: string; csrfToken?: string }>(
-        "/api/auth/refresh",
-        undefined,
-        { skipAuthRefresh: true } as any,
-      )
-      .then(({ data }) => {
-        setJwtToken(data.token);
-        // A token rotation also rotates the CSRF pair; pick up the new one
-        // so the next mutation doesn't hit a stale-cookie 403.
-        if (data.csrfToken) csrfToken = data.csrfToken;
-        return data.token;
-      })
-      .catch((error) => {
-        setJwtToken(null);
-        throw error;
-      })
-      .finally(() => {
-        refreshPromise = null;
-      });
-  }
-
-  return refreshPromise;
-}
-
-export async function logout() {
-  try {
-    await api.post("/api/auth/logout", undefined, { skipAuthRefresh: true } as any);
-  } finally {
-    setJwtToken(null);
-    clearCsrfToken();
-  }
-}
-
-// ─── Categories ────────────────────────────────────────────────────────────────
-
-export interface CategoryNode {
-  id: number;
-  slug: string;
-  name: string;
-  children: CategoryNode[];
-}
-
-export async function fetchCategories(): Promise<CategoryNode[]> {
-  try {
-    const { data } = await api.get<{ success: boolean; data: CategoryNode[] }>("/api/categories");
-    return data.data;
-  } catch {
-    return [];
-  }
-}
-
-// ─── Jobs ─────────────────────────────────────────────────────────────────────
-
-export async function fetchJobs(params?: {
-  category?: string;
-  status?: string;
-  limit?: number;
-  search?: string;
-  after?: string;
-  cursor?: string;
-  timezone?: string;
-  viewerAddress?: string;
-  minBudget?: string;
-  maxBudget?: string;
-  skills?: string;
-  minClientRating?: string;
-  duration?: string;
-  postedSince?: string;
-  maxApplications?: string;
-}) {
-  const {
-    minBudget,
-    maxBudget,
-    minClientRating,
-    postedSince,
-    maxApplications,
-    after,
-    ...rest
-  } = params || {};
-
-  const { data } = await api.get<{
-    success: boolean;
-    data: Job[];
-    next_cursor: string | null;
-    has_more: boolean;
-  }>("/api/jobs", {
-    params: {
-      ...rest,
-      after,
-      min_budget: minBudget,
-      max_budget: maxBudget,
-      skills: params?.skills,
-      min_client_rating: minClientRating,
-      duration: params?.duration,
-      posted_since: postedSince,
-      max_applications: maxApplications,
-    },
-  });
-
-  return {
-    jobs: data.data,
-    nextCursor: data.next_cursor ?? null,
-    hasMore: data.has_more ?? Boolean(data.next_cursor),
-  };
-}
-
-export interface JobSuggestion {
-  type: "title" | "skill" | "category";
-  value: string;
-}
-
-export async function fetchJobSuggestions(query: string): Promise<JobSuggestion[]> {
-  try {
-    const { data } = await api.get<{ success: boolean; data: JobSuggestion[] }>(
-      "/api/jobs/suggestions",
-      { params: { q: query } },
-    );
-    return data.data;
-  } catch {
-    return [];
-  }
-}
-
-export async function fetchRelatedJobs(category: string, currentJobId: string) {
-  const { jobs } = await fetchJobs({
-    category,
-    status: "open",
-    limit: 4,
-  });
-
-  return jobs.filter((job) => job.id !== currentJobId).slice(0, 3);
-}
-
-export async function fetchRecentlyCompletedJobs(limit = 3): Promise<Job[]> {
-  const { jobs } = await fetchJobs({ status: "completed", limit });
-  return jobs;
-}
-
-export interface InsightCategory {
-  category: string;
-  totalJobs: number;
-  avgBudget: number;
-  avgApplicationsPerJob: number;
-  acceptanceRate: number;
-  lowCompetitionJobs: number;
-  uniqueClients: number;
-}
-
-export interface InsightClientMix {
-  newClients: number;
-  returningClients: number;
-  totalClients: number;
-}
-
-export interface InsightSkill {
-  skill: string;
-  demandCount: number;
-  avgApplicationsPerJob: number;
-  lowCompetitionJobs: number;
-}
-
-export interface InsightCompetitiveJob {
-  id: string;
-  title: string;
-  category: string;
-  budget: number;
-  currency: string;
-  clientAddress: string;
-  createdAt: string;
-  applicationCount: number;
-  competitionLevel: "uncontested" | "light" | "active";
-}
-
-export interface InsightPayTrend {
-  date: string;
-  category: string;
-  avgBudget: number;
-  jobCount: number;
-}
-
-export async function fetchInsightCategories(limit = 20) {
-  const { data } = await api.get<{
-    success: boolean;
-    data: {
-      categories: InsightCategory[];
-      clientMix: InsightClientMix;
-    };
-  }>("/api/insights/categories", { params: { limit } });
-  return data.data;
-}
-
-export async function fetchInsightSkills(limit = 20) {
-  const { data } = await api.get<{ success: boolean; data: InsightSkill[] }>(
-    "/api/insights/skills",
-    { params: { limit } },
-  );
-  return data.data;
-}
-
-export async function fetchInsightCompetitive(limit = 20) {
-  const { data } = await api.get<{ success: boolean; data: InsightCompetitiveJob[] }>(
-    "/api/insights/competitive",
-    { params: { limit } },
-  );
-  return data.data;
-}
-
-export async function fetchInsightPayTrends(days = 30) {
-  const { data } = await api.get<{ success: boolean; data: InsightPayTrend[] }>(
-    "/api/insights/trends/pay",
-    { params: { days } },
-  );
-  return data.data;
-}
-
-/**
- * Fetches a single job by its identifier.
- *
- * @param id Job identifier.
- * @returns The matching job record.
- * @throws {import("axios").AxiosError} If the job is not found or the request fails.
- * @see backend/src/routes/jobs.js
- */
-export async function fetchJob(id: string, viewerAddress?: string) {
-  const { data } = await api.get<{ success: boolean; data: Job }>(
-    `/api/jobs/${id}`,
-    {
-      params: viewerAddress ? { viewerAddress } : undefined,
-    },
-  );
-  return data.data;
-}
-
-export async function createJob(payload: {
-  title: string;
-  description: string;
-  budget: string;
-  currency?: "XLM" | "USDC";
-  category: string;
-  skills: string[];
-  deadline?: string;
-  timezone?: string;
-  clientAddress: string;
-  screeningQuestions?: string[];
-  visibility?: "public" | "private" | "invite_only";
-  milestones?: { description: string; amount: string }[];
-}) {
-  const { data } = await api.post<{ success: boolean; data: Job }>(
-    "/api/jobs",
-    payload,
-  );
-  return data.data;
-}
-
-export async function fetchMyJobs(publicKey: string) {
-  const { data } = await api.get<{ success: boolean; data: Job[] }>(
-    `/api/jobs/client/${publicKey}`,
-  );
-  return data.data;
-}
-
-/**
- * Evaluates application quality using AI (Claude API).
- *
- * @param jobId Job identifier.
- * @returns Array of scores and reasonings for all applications.
- */
-export async function scoreProposals(jobId: string) {
-  const { data } = await api.post<{
-    success: boolean;
-    data: { id: string; score: number; reasoning: string }[];
-  }>(`/api/jobs/${jobId}/score-proposals`);
-  return data.data;
-}
-
-/**
- * Get analytics for a job (applications per day, avg bid, skill distribution, time to hire).
- *
- * @param jobId Job identifier.
- * @returns Analytics data for the job.
- */
-export async function fetchJobAnalytics(jobId: string) {
-  const { data } = await api.get<{ success: boolean; data: JobAnalytics }>(
-    `/api/jobs/${jobId}/analytics`,
-  );
-  return data.data;
-}
-
-/**
- * Extend a job's expiry by the given number of days.
- * Charges a 0.5 XLM fee per 7-day block.
- *
- * @param jobId Job identifier.
- * @param days Number of days to extend (7, 14, or 30).
- * @returns Updated job record.
- */
-export async function extendJobExpiry(jobId: string, days = 30) {
-  const { data } = await api.patch<{ success: boolean; data: Job }>(
-    `/api/jobs/${jobId}/extend`,
-    { days },
-  );
-  return data.data;
-}
-
-/**
- * Get jobs expiring within 3 days.
- *
- * @returns Array of expiring jobs.
- */
-export async function fetchExpiringJobs() {
-  const { data } = await api.get<{ success: boolean; data: Job[] }>(
-    "/api/jobs/expiring",
-  );
-  return data.data;
-}
-
-/**
- * Manually trigger expiry check for old jobs.
- *
- * @returns Count of expired jobs.
- */
-export async function expireOldJobs() {
-  const { data } = await api.post<{
-    success: boolean;
-    data: { expiredCount: number };
-  }>("/api/jobs/expire-old");
-  return data.data.expiredCount;
-}
-
-// ─── Applications ─────────────────────────────────────────────────────────────
-
-export async function fetchApplications(jobId: string, tier?: string) {
-  const { data } = await api.get<{ success: boolean; data: Application[] }>(
-    `/api/applications/job/${jobId}`,
-    { params: tier ? { tier } : undefined },
-  );
-  return data.data;
-}
-
-export async function submitApplication(payload: {
-  jobId: string;
-  freelancerAddress: string;
-  proposal: string;
-  bidAmount: string;
-  currency: string;
-  bidCommitment?: string;
-  bidNonce?: string;
-  screeningAnswers?: Record<string, string>;
-  referredBy?: string;
-}) {
-  const { data } = await api.post<{ success: boolean; data: Application }>(
-    "/api/applications",
-    payload,
-  );
-  return data.data;
-}
-
-export async function closeBidding(jobId: string, clientAddress: string) {
-  const { data } = await api.post(`/api/applications/job/${jobId}/close-bidding`, {
-    clientAddress,
-  });
-  return data.data;
-}
-
-export async function revealApplicationBid(
-  applicationId: string,
-  payload: { freelancerAddress: string; bidAmount: string; nonce: string },
-) {
-  const { data } = await api.post(`/api/applications/${applicationId}/reveal`, payload);
-  return data.data;
-}
-
-export async function acceptApplication(
-  applicationId: string,
-  clientAddress: string,
-) {
-  const { data } = await api.post(`/api/applications/${applicationId}/accept`, {
-    clientAddress,
-  });
-  return data.data;
-}
-
-export async function fetchMyApplications(publicKey: string) {
-  const { data } = await api.get<{ success: boolean; data: Application[] }>(
-    `/api/applications/freelancer/${publicKey}`,
-  );
-  return data.data;
-}
-
-// ─── Profile ──────────────────────────────────────────────────────────────────
-
-export async function fetchProfile(publicKey: string) {
-  const { data } = await api.get<{ success: boolean; data: UserProfile }>(
-    `/api/profiles/${publicKey}`,
-  );
-  return data.data;
-}
-
-export async function fetchProfileResponseTime(publicKey: string) {
-  const { data } = await api.get<{
-    success: boolean;
-    data: { averageDays: number | null };
-  }>(`/api/profiles/${encodeURIComponent(publicKey)}/response-time`);
-  return data.data;
-}
-
-export async function fetchPublicProfile(
-  publicKey: string,
-): Promise<UserProfile | null> {
-  try {
-    const { data } = await api.get<{ success: boolean; data: UserProfile }>(
-      `/api/profiles/${encodeURIComponent(publicKey)}`,
-    );
-    return data.data;
-  } catch (e) {
-    if (axios.isAxiosError(e) && e.response?.status === 404) return null;
-    throw e;
-  }
-}
-
-export async function fetchProfiles(params?: {
-  role?: string;
-  availability?: string;
-  search?: string;
-  limit?: number;
-  after?: string;
-}) {
-  const { data } = await api.get<{
-    success: boolean;
-    data: UserProfile[];
-    next_cursor: string | null;
-    has_more: boolean;
-  }>("/api/profiles", { params });
-  return {
-    profiles: data.data,
-    nextCursor: data.next_cursor ?? null,
-    hasMore: data.has_more ?? false,
-  };
-}
-
-export async function syncOnboardingProgress(payload: {
-  publicKey: string;
-  currentStep: number;
-  completedSteps: string[];
-  dismissed: boolean;
-  completed: boolean;
-}) {
-  const { data } = await api.patch<{ success: boolean; data?: unknown }>(
-    "/api/onboarding",
-    payload,
-  );
-  return data;
-}
-
-export async function searchFreelancers(params?: { search?: string; limit?: number }) {
-  const { data } = await api.get<{ success: boolean; data: UserProfile[] }>(
-    "/api/freelancers",
-    { params },
-  );
-  return data.data;
-}
-
-export async function fetchProfileStats(publicKey: string): Promise<ProfileStats> {
-  const { data } = await api.get<{ success: boolean; data: ProfileStats }>(
-    `/api/profiles/${encodeURIComponent(publicKey)}/stats`,
-  );
-  return data.data;
-}
-
-export async function fetchResponseTime(publicKey: string): Promise<ResponseTime> {
-  const { data } = await api.get<{ success: boolean; data: ResponseTime }>(
-    `/api/profiles/${encodeURIComponent(publicKey)}/response-time`,
-  );
-  return data.data;
-}
-
-export async function upsertProfile(
-  payload: Partial<UserProfile> & { publicKey: string },
-) {
-  const { data } = await api.post<{ success: boolean; data: UserProfile }>(
-    "/api/profiles",
-    payload,
-  );
-  return data.data;
-}
-
-export async function updateProfileAvailability(
-  publicKey: string,
-  payload: Availability,
-) {
-  const { data } = await api.post<{ success: boolean; data: UserProfile }>(
-    `/api/profiles/${encodeURIComponent(publicKey)}/availability`,
-    payload,
-  );
-  return data.data;
-}
-
-/**
- * Verifies a user's identity via a DID provider and stores the resulting credential hash.
- *
- * @param publicKey User Stellar public key.
- * @param didHash The credential hash/DID URI returned by the provider.
- * @returns The updated profile.
- */
-export async function verifyIdentity(publicKey: string, didHash: string) {
-  const { data } = await api.post<{ success: boolean; data: UserProfile }>(
-    `/api/profiles/${encodeURIComponent(publicKey)}/verify`,
-    { didHash },
-  );
-  return data.data;
-}
-
-// ─── Escrow ───────────────────────────────────────────────────────────────────
-
-export async function fetchEscrow(jobId: string) {
-  const { data } = await api.get<{ success: boolean; data: any }>(
-    `/api/escrow/${jobId}`,
-  );
-  return data.data;
-}
-
-export async function releaseEscrow(
-  jobId: string,
-  clientAddress: string,
-  contractTxHash?: string,
-  releaseCurrency?: "XLM" | "USDC",
-) {
-  const { data } = await api.post(`/api/escrow/${jobId}/release`, {
-    clientAddress,
-    ...(contractTxHash ? { contractTxHash } : {}),
-    ...(releaseCurrency ? { releaseCurrency } : {}),
-  });
-  return data.data;
-}
-
-export async function releaseMilestone(
-  jobId: string,
-  clientAddress: string,
-  milestoneIndex: number,
-  contractTxHash?: string,
-) {
-  const { data } = await api.post(`/api/escrow/${jobId}/release-milestone`, {
-    clientAddress,
-    milestoneIndex,
-    ...(contractTxHash ? { contractTxHash } : {}),
-  });
-  return data.data;
-}
-
-export async function rejectMilestone(
-  jobId: string,
-  clientAddress: string,
-  milestoneIndex: number,
-  contractTxHash?: string,
-) {
-  const { data } = await api.post(`/api/escrow/${jobId}/reject-milestone`, {
-    clientAddress,
-    milestoneIndex,
-    ...(contractTxHash ? { contractTxHash } : {}),
-  });
-  return data.data;
-}
-
-export async function disputeMilestone(
-  jobId: string,
-  raisedBy: string,
-  milestoneIndex: number,
-) {
-  const { data } = await api.post(`/api/escrow/${jobId}/dispute-milestone`, {
-    raisedBy,
-    milestoneIndex,
-  });
-  return data.data;
-}
-
-export async function timeoutRefund(
-  jobId: string,
-  clientAddress: string,
-  contractTxHash?: string,
-) {
-  const { data } = await api.post(`/api/escrow/${jobId}/timeout-refund`, {
-    clientAddress,
-    ...(contractTxHash ? { contractTxHash } : {}),
-  });
-  return data.data;
-}
-
-export async function inviteFreelancer(
-  jobId: string,
-  freelancerAddress: string,
-) {
-  const { data } = await api.post<{ success: boolean; data: any }>(
-    `/api/jobs/${jobId}/invite`,
-    {
-      freelancerAddress,
-    },
-  );
-  return data.data;
-}
-
-export async function fetchProposalTemplates() {
-  const { data } = await api.get<{
-    success: boolean;
-    data: ProposalTemplate[];
-  }>("/api/proposal-templates");
-  return data.data;
-}
-
-export async function createProposalTemplate(payload: {
-  name: string;
-  content: string;
-}) {
-  const { data } = await api.post<{ success: boolean; data: ProposalTemplate }>(
-    "/api/proposal-templates",
-    payload,
-  );
-  return data.data;
-}
-
-export async function updateProposalTemplate(
-  id: string,
-  payload: { name?: string; content?: string },
-) {
-  const { data } = await api.patch<{
-    success: boolean;
-    data: ProposalTemplate;
-  }>(`/api/proposal-templates/${id}`, payload);
-  return data.data;
-}
-
-export async function deleteProposalTemplate(id: string) {
-  await api.delete(`/api/proposal-templates/${id}`);
-}
-
-// ─── Saved Searches ─────────────────────────────────────────────────────────────
-
-export interface SavedSearch {
-  id: string;
-  user_address: string;
-  query_params: Record<string, any>;
-  notify_in_app: boolean;
-  notify_email: boolean;
-  last_notified_at: string | null;
-  created_at: string;
-  updated_at: string;
-}
-
-export async function fetchSavedSearches(): Promise<SavedSearch[]> {
-  const { data } = await api.get<{ success: boolean; data: SavedSearch[] }>(
-    "/api/saved-searches",
-  );
-  return data.data;
-}
-
-export async function createSavedSearch(payload: {
-  query_params: Record<string, any>;
-  notify_in_app?: boolean;
-  notify_email?: boolean;
-}): Promise<SavedSearch> {
-  const { data } = await api.post<{ success: boolean; data: SavedSearch }>(
-    "/api/saved-searches",
-    payload,
-  );
-  return data.data;
-}
-
-export async function updateSavedSearch(
-  id: string,
-  payload: {
-    notify_in_app?: boolean;
-    notify_email?: boolean;
-  },
-): Promise<SavedSearch> {
-  const { data } = await api.patch<{ success: boolean; data: SavedSearch }>(
-    `/api/saved-searches/${id}`,
-    payload,
-  );
-  return data.data;
-}
-
-export async function deleteSavedSearch(id: string): Promise<void> {
-  await api.delete(`/api/saved-searches/${id}`);
-}
-
-export async function fetchPriceAlertPreference(publicKey: string) {
-  const { data } = await api.get<{
-    success: boolean;
-    data: PriceAlertPreference | null;
-  }>(`/api/profiles/${encodeURIComponent(publicKey)}/price-alerts`);
-  return data.data;
-}
-
-export async function fetchClientSpendingAnalytics(publicKey: string) {
-  const { data } = await api.get<{
-    success: boolean;
-    data: ClientSpendingAnalytics;
-  }>(`/api/profiles/${encodeURIComponent(publicKey)}/spending`);
-  return data.data;
-}
-
-export async function fetchClientReputation(publicKey: string): Promise<ClientReputation> {
-  const { data } = await api.get<{ success: boolean; data: ClientReputation }>(
-    `/api/profiles/${encodeURIComponent(publicKey)}/client-reputation`
-  );
-  return data.data;
-}
-
-export interface XlmPriceHistoryPoint {
-  timestamp: number;
-  priceUsd: number;
-}
-
-export interface XlmPriceHistory {
-  points: XlmPriceHistoryPoint[];
-  currentPriceUsd: number | null;
-  change24hPercent: number | null;
-  updatedAt?: string;
-  cached?: boolean;
-}
-
-export type Timeframe = '1D' | '7D' | '30D';
-
-const HORIZON_URL =
-  process.env.NEXT_PUBLIC_HORIZON_URL || "https://horizon-testnet.stellar.org";
-
-const NETWORK = (process.env.NEXT_PUBLIC_STELLAR_NETWORK || "testnet") as "testnet" | "mainnet";
-
-const USDC_ISSUER =
-  NETWORK === "mainnet"
-    ? "GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN"
-    : "GBBD47IF6LWK7P7MDEVSCWR7DPUWV3NY3DTQEVFL4NAT4AQH3ZLLFLA5";
-
-const TIMEFRAME_CONFIG: Record<Timeframe, { resolutionMs: number; lookbackMs: number }> = {
-  '1D':  { resolutionMs: 3_600_000,  lookbackMs: 24 * 60 * 60 * 1000 },
-  '7D':  { resolutionMs: 86_400_000, lookbackMs: 7  * 24 * 60 * 60 * 1000 },
-  '30D': { resolutionMs: 86_400_000, lookbackMs: 30 * 24 * 60 * 60 * 1000 },
-};
-
-export async function fetchXlmPriceHistory(timeframe: Timeframe = '7D'): Promise<XlmPriceHistory> {
-  const { resolutionMs, lookbackMs } = TIMEFRAME_CONFIG[timeframe];
-  const nowMs = Date.now();
-  const startMs = nowMs - lookbackMs;
-
-  const url = new URL(`${HORIZON_URL}/trade_aggregations`);
-  url.searchParams.set('base_asset_type',      'native');
-  url.searchParams.set('counter_asset_code',   'USDC');
-  url.searchParams.set('counter_asset_issuer', USDC_ISSUER);
-  url.searchParams.set('counter_asset_type',   'credit_alphanum4');
-  url.searchParams.set('resolution',           String(resolutionMs));
-  url.searchParams.set('start_time',           String(startMs));
-  url.searchParams.set('end_time',             String(nowMs));
-  url.searchParams.set('order',                'asc');
-  url.searchParams.set('limit',                '200');
-
-  const response = await fetch(url.toString());
-
-  if (!response.ok) {
-    throw new Error(`Horizon API error: ${response.status}`);
-  }
-
-  const json = await response.json();
-  const records: { timestamp: string; close: string }[] =
-    json?._embedded?.records ?? [];
-
-  const points: XlmPriceHistoryPoint[] = records.map((r) => ({
-    timestamp: Number(r.timestamp),
-    priceUsd:  parseFloat(r.close),
-  }));
-
-  const last  = points.length > 0 ? points[points.length - 1] : null;
-  const first = points.length > 0 ? points[0] : null;
-
-  const currentPriceUsd = last ? last.priceUsd : null;
-  const change24hPercent =
-    points.length >= 2 && first && last
-      ? ((last.priceUsd - first.priceUsd) / first.priceUsd) * 100
-      : null;
-
-  return { points, currentPriceUsd, change24hPercent };
-}
-
-export async function upsertPriceAlertPreference(
-  publicKey: string,
-  payload: {
-    minXlmPriceUsd?: number | null;
-    maxXlmPriceUsd?: number | null;
-    emailNotificationsEnabled?: boolean;
-    email?: string;
-  },
-) {
-  const { data } = await api.post<{
-    success: boolean;
-    data: PriceAlertPreference;
-  }>(`/api/profiles/${encodeURIComponent(publicKey)}/price-alerts`, payload);
-  return data.data;
-}
-
-/**
- * Stores the on-chain escrow contract ID against a job record.
- *
- * @param jobId Job identifier.
- * @param escrowContractId Soroban transaction hash returned after create_escrow().
- * @returns The updated job record.
- */
-export async function updateJobEscrowId(
-  jobId: string,
-  escrowContractId: string,
-) {
-  const { data } = await api.patch<{ success: boolean; data: Job }>(
-    `/api/jobs/${jobId}/escrow`,
-    { escrowContractId },
-  );
-  return data.data;
-}
-
-export async function deleteJob(jobId: string) {
-  await api.delete(`/api/jobs/${jobId}`);
-}
-
-/**
- * Raises a dispute for an in-progress job.
- *
- * @param jobId Job identifier.
- * @param payload Dispute details (reason and description).
- * @returns The updated job record.
- */
-export async function raiseDispute(
-  jobId: string,
-  payload: { reason: string; description: string },
-) {
-  const { data } = await api.post<{ success: boolean; data: Job }>(
-    `/api/jobs/${jobId}/dispute`,
-    payload,
-  );
-  return data.data;
-}
-
-/**
- * Resolves a dispute for a job (Admin only).
- *
- * @param jobId Job identifier.
- * @param note Resolution note.
- * @param releaseTo Release funds to "client" or "freelancer".
- * @returns The updated job record.
- */
-export async function resolveDispute(jobId: string, note?: string, releaseTo?: string) {
-  const { data } = await api.post<{ success: boolean; data: Job }>(
-    `/api/jobs/${jobId}/resolve`,
-    { note, releaseTo },
-  );
-  return data.data;
-}
-
-// ─── Time entries ─────────────────────────────────────────────────────────────
-
-export async function logTimeEntry(payload: {
-  jobId: string;
-  durationMinutes: number;
-  description?: string;
-  startedAt?: string;
-}) {
-  const { data } = await api.post<{ success: boolean; data: TimeEntry }>(
-    "/api/time-entries",
-    payload,
-  );
-  return data.data;
-}
-
-export async function fetchTimeEntries(jobId: string): Promise<TimeEntry[]> {
-  const { data } = await api.get<{ success: boolean; data: TimeEntry[] }>(
-    `/api/time-entries/job/${jobId}`,
-  );
-  return data.data;
-}
-
-export async function fetchTimeInvoices(jobId: string): Promise<TimeInvoice[]> {
-  const { data } = await api.get<{ success: boolean; data: TimeInvoice[] }>(
-    `/api/time-entries/job/${jobId}/invoices`,
-  );
-  return data.data;
-}
-
-export async function generateTimeInvoice(payload: {
-  jobId: string;
-  hourlyRateXlm: number;
-}) {
-  const { data } = await api.post<{ success: boolean; data: TimeInvoice }>(
-    "/api/time-entries/invoice",
-    payload,
-  );
-  return data.data;
-}
-
-export async function reviewTimeInvoice(
-  invoiceId: string,
-  decision: "approved" | "rejected",
-) {
-  const { data } = await api.patch<{ success: boolean; data: TimeInvoice }>(
-    `/api/time-entries/invoice/${invoiceId}/review`,
-    { decision },
-  );
-  return data.data;
-}
-
-// ─── Ratings ──────────────────────────────────────────────────────────────────
-
-export async function submitRating(payload: {
-  jobId: string;
-  ratedAddress: string;
-  stars: number;
-  review?: string;
-}) {
-  const { data } = await api.post<{ success: boolean; data: Rating }>(
-    "/api/ratings",
-    payload,
-  );
-  return data.data;
-}
-
-export async function fetchRatings(publicKey: string) {
-  const { data } = await api.get<{ success: boolean; data: Rating[] }>(
-    `/api/ratings/${publicKey}`,
-  );
-  return data.data;
-}
-
-// ─── Recommendations ──────────────────────────────────────────────────────────
-
-export async function fetchRecommendedJobs(
-  publicKey: string,
-): Promise<(Job & { matchScore: number })[]> {
-  const { data } = await api.get<{
-    success: boolean;
-    data: (Job & { matchScore: number })[];
-  }>(`/api/jobs/recommended/${encodeURIComponent(publicKey)}`);
-  return data.data;
-}
-
-export async function fetchDrafts() {
-  const { data } = await api.get<{ success: boolean; data: any[] }>(
-    "/api/jobs/drafts",
-  );
-  return data.data;
-}
-
-export async function fetchDraft(draftId: string) {
-  const { data } = await api.get<{ success: boolean; data: any }>(
-    `/api/jobs/drafts/${draftId}`,
-  );
-  return data.data;
-}
-
-export async function saveDraft(draft: {
-  id?: string;
-  title?: string;
-  description?: string;
-  budget?: number;
-  category?: string;
-  skills?: string[];
-  deadline?: string;
-}) {
-  const { data } = await api.post<{ success: boolean; data: { id: string } }>("/api/jobs/drafts", draft);
-  return data.data;
-}
-
-export async function updateDraft(draft: {
-  id: string;
-  title?: string;
-  description?: string;
-  budget?: number;
-  category?: string;
-  skills?: string[];
-  deadline?: string;
-}) {
-  const { data } = await api.put<{ success: boolean; data: { id: string } }>(`/api/jobs/drafts/${draft.id}`, draft);
-  return data.data;
-}
-
-export async function deleteDraft(draftId: string) {
-  await api.delete(`/api/jobs/drafts/${draftId}`);
-}
-
-// ─── Skill Assessments ─────────────────────────────────────────────────────────
-
-export async function fetchAssessment(skill: string) {
-  const { data } = await api.get<{
-    success: boolean;
-    data: {
-      label: string;
-      skill: string;
-      questions: AssessmentQuestion[];
-      durationSeconds: number;
-      canRetake: boolean;
-      retakeAvailableAt?: string;
-      lastAttempt?: { score: number; passed: boolean };
-    };
-  }>(`/api/assessments/${encodeURIComponent(skill)}`);
-  return data.data;
-}
-
-export async function submitAssessment(
-  skill: string,
-  answers: Record<number, number>,
-) {
-  const { data } = await api.post<{
-    success: boolean;
-    data: {
-      score: number;
-      passed: boolean;
-      correct: number;
-      total: number;
-    };
-  }>(`/api/assessments/${encodeURIComponent(skill)}/submit`, { answers });
-  return data.data;
-}
-
-// ─── Admin 2FA ────────────────────────────────────────────────────────────────
-
-export async function fetchAdmin2FAStatus() {
-  const { data } = await api.get<{
-    success: boolean;
-    data: { totp_enabled: boolean; verified: boolean };
-  }>("/api/admin/2fa/status");
-  return data.data;
-}
-
-export async function setupAdmin2FA() {
-  const { data } = await api.post<{
-    success: boolean;
-    data: { qrCode: string; manualEntryKey: string };
-  }>("/api/admin/2fa/setup");
-  return data.data;
-}
-
-export async function verifyAdmin2FA(token: string, setup = false) {
-  const { data } = await api.post<{
-    success: boolean;
-    token?: string;
-    data: { backupCodes?: string[]; message?: string };
-  }>("/api/admin/2fa/verify", { token, setup });
-  return { token: data.token, backupCodes: data.data?.backupCodes, message: data.data?.message };
-}
-
-// ─── Bulk Job Actions ───────────────────────────────────────────────────────
-
-export async function bulkCancelJobs(jobIds: string[]): Promise<BulkActionResponse> {
-  const { data } = await api.post<{ success: boolean; data: BulkActionResponse }>(
-    "/api/jobs/bulk-cancel",
-    { jobIds },
-  );
-  return data.data;
-}
-
-export async function bulkExtendJobs(jobIds: string[], days: number): Promise<BulkActionResponse> {
-  const { data } = await api.post<{ success: boolean; data: BulkActionResponse }>(
-    "/api/jobs/bulk-extend",
-    { jobIds, days },
-  );
-  return data.data;
-}
-
-export async function bulkBoostJobs(jobIds: string[], txHash: string): Promise<BulkActionResponse> {
-  const { data } = await api.post<{ success: boolean; data: BulkActionResponse }>(
-    "/api/jobs/bulk-boost",
-    { jobIds, txHash },
-  );
-  return data.data;
-}
-
-
-// ─── IPFS File Upload (Issue #202) ──────────────────────────────────────────
-
-export async function uploadPortfolioFiles(
-  publicKey: string,
-  files: FileList | File[],
-  onProgress?: (fileIndex: number, percent: number) => void,
-) {
-  const formData = new FormData();
-  const filesArr = Array.from(files);
-  filesArr.forEach((file) => {
-    formData.append("files", file);
-  });
-
-  const { data } = await api.post<{
-    success: boolean;
-    data: {
-      uploadedFiles: PortfolioFile[];
-      gatewayUrls: string[];
-    };
-  }>(`/api/profiles/${encodeURIComponent(publicKey)}/portfolio-files`, formData, {
-    headers: {
-      "Content-Type": "multipart/form-data",
-    },
-    timeout: 120000,
-    onUploadProgress: (e) => {
-      if (onProgress && e.total) {
-        const overallPercent = Math.round((e.loaded / e.total) * 100);
-        filesArr.forEach((_, i) => onProgress(i, overallPercent));
-      }
-    },
-  });
-
-  return data.data;
-}
-
-// ─── Stellar Faucet (Issue #205) ───────────────────────────────────────────
-
-export async function fundTestnetWallet(publicKey: string) {
-  const { data } = await api.post<{
-    success: boolean;
-    data: {
-      success: boolean;
-      message: string;
-      fundedAmount: string;
-      newBalance?: string;
-      transactionHash?: string;
-      ledger?: number;
-    };
-  }>("/api/faucet/fund", { publicKey });
-
-  return data.data;
-}
-
-export async function checkAccountNeedsFunding(publicKey: string) {
-  const { data } = await api.get<{
-    success: boolean;
-    data: {
-      needsFunding: boolean;
-      currentBalance: string;
-      exists: boolean;
-    };
-  }>(`/api/faucet/check/${encodeURIComponent(publicKey)}`);
-
-  return data.data;
-}
-
-export async function getFaucetStatus() {
-  const { data } = await api.get<{
-    success: boolean;
-    data: {
-      enabled: boolean;
-      network: string;
-      amount: string;
-      asset: string;
-    };
-  }>("/api/faucet/status");
-
-  return data.data;
-}
-
-// ─── Token Support (Issue #228) ─────────────────────────────────────────────
-
-export async function getPopularTokens() {
-  const { data } = await api.get<{
-    success: boolean;
-    data: TokenInfo[];
-  }>("/api/tokens/popular");
-
-  return data.data;
-}
-
-export async function searchTokens(query: string) {
-  const { data } = await api.get<{
-    success: boolean;
-    data: TokenInfo[];
-  }>("/api/tokens/search", { params: { q: query } });
-
-  return data.data;
-}
-
-export async function getTokenMetadata(contractId: string) {
-  const { data } = await api.get<{
-    success: boolean;
-    data: TokenInfo;
-  }>(`/api/tokens/${contractId}/metadata`);
-
-  return data.data;
-}
-
-export async function getTokenBalance(contractId: string, publicKey: string) {
-  const { data } = await api.get<{
-    success: boolean;
-    data: TokenBalance;
-  }>(`/api/tokens/${contractId}/balance/${publicKey}`);
-
-  return data.data;
-}
-
-export async function validateTokenContract(contractId: string) {
-  const { data } = await api.post<{
-    success: boolean;
-    data: {
-      valid: boolean;
-      error?: string;
-    };
-  }>("/api/tokens/validate", { contractId });
-
-  return data.data;
-}
-
-// ─── Stellar Turrets (Issue #224) ───────────────────────────────────────────
-
-export async function submitViaTurrets(
-  transactionXDR: string,
-  useTurret?: boolean,
-) {
-  const { data } = await api.post<{
-    success: boolean;
-    data: {
-      success: boolean;
-      hash: string;
-      ledger: number;
-      feeCharged: string;
-      turretUsed: boolean;
-      message: string;
-    };
-  }>("/api/turrets/submit", { transactionXDR, useTurret });
-
-  return data.data;
-}
-
-export async function getTurretsStatus() {
-  const { data } = await api.get<{
-    success: boolean;
-    data: {
-      available: boolean;
-      url?: string;
-      network?: string;
-      version?: string;
-      feeSponsorship?: boolean;
-      message: string;
-      error?: string;
-    };
-  }>("/api/turrets/status");
-
-  return data.data;
-}
-
-export async function estimateTurretsFee(transactionXDR: string) {
-  const { data } = await api.post<{
-    success: boolean;
-    data: {
-      success: boolean;
-      baseFee: string;
-      turretFee: string;
-      totalFee: string;
-      feeSponsored: boolean;
-      message?: string;
-    };
-  }>("/api/turrets/estimate", { transactionXDR });
-
-  return data.data;
-}
-
-export async function getTurretsConfig() {
-  const { data } = await api.get<{
-    success: boolean;
-    data: {
-      configured: boolean;
-      url: string | null;
-      hasApiKey: boolean;
-      shouldUseByDefault: boolean;
-    };
-  }>("/api/turrets/config");
-
-  return data.data;
-}
-
-// ─── Messages ──────────────────────────────────────────────────────────────────
-
-/**
- * Fetches all messages for a specific job.
- * Automatically marks messages as read for the current user.
- *
- * @param jobId Job identifier.
- * @returns Messages sorted chronologically (oldest first).
- * @throws {import("axios").AxiosError} If unauthorized, job not found, or request fails.
- * @see backend/src/routes/messageRoutes.js
- */
-export async function fetchMessages(jobId: string): Promise<Message[]> {
-  const { data } = await api.get<{ success: boolean; data: Message[] }>(
-    `/api/messages/job/${jobId}`,
-  );
-  return data.data;
-}
-
-/**
- * Sends a message in a job thread.
- *
- * Request payload shape:
- * - `content` (string): message text (1-2000 characters).
- *
- * @param jobId Job identifier.
- * @param content Message content.
- * @returns The created message object.
- * @throws {import("axios").AxiosError} If unauthorized, validation fails, or request fails.
- * @see backend/src/routes/messageRoutes.js
- */
-export async function sendMessage(
-  jobId: string,
-  content: string,
-  contractTxHash?: string,
-): Promise<Message> {
-  const { data } = await api.post<{ success: boolean; data: Message }>(
-    `/api/messages/job/${jobId}`,
-    { content, contractTxHash },
-  );
-  return data.data;
-}
-
-/**
- * Fetches the total unread message count for the authenticated user.
- *
- * @returns Number of unread messages.
- * @throws {import("axios").AxiosError} If not authenticated or request fails.
- * @see backend/src/routes/messageRoutes.js
- */
-export async function fetchUnreadCount(): Promise<number> {
-  const { data } = await api.get<{
-    success: boolean;
-    data: { unreadCount: number };
-  }>("/api/messages/unread-count");
-  return data.data.unreadCount;
-}
-
-export interface NotificationsResponse {
-  notifications: NotificationItem[];
-  unreadCount: number;
-  nextCursor: string | null;
-  hasMore: boolean;
-}
-
-export async function fetchNotifications(params?: {
-  limit?: number;
-  after?: string;
-  cursor?: string | null;
-}): Promise<NotificationsResponse> {
-  const { data } = await api.get<{
-    success: boolean;
-    data: NotificationsResponse;
-  }>("/api/notifications", {
-    params: {
-      limit: params?.limit,
-      after: params?.after,
-      cursor: params?.cursor || undefined,
-    },
-  });
-  return {
-    ...data.data,
-    hasMore: data.data.has_more ?? Boolean(data.data.next_cursor),
-  };
-}
-
-export async function markNotificationRead(id: string): Promise<NotificationItem> {
-  const { data } = await api.patch<{
-    success: boolean;
-    data: NotificationItem;
-  }>(`/api/notifications/${id}/read`);
-  return data.data;
-}
-
-export async function markAllNotificationsRead(): Promise<{ updatedCount: number }> {
-  const { data } = await api.patch<{
-    success: boolean;
-    data: { updatedCount: number };
-  }>("/api/notifications/read-all");
-  return data.data;
-}
-
-/**
- * Attaches an on-chain Soroban transaction hash to a message record.
- * Called after the frontend signs and submits the publish_message event.
- */
-export async function attachMessageTxHash(
-  messageId: string,
-  txHash: string,
-): Promise<Message> {
-  const { data } = await api.patch<{ success: boolean; data: Message }>(
-    `/api/messages/${messageId}/tx-hash`,
-    { txHash },
-  );
-  return data.data;
-}
-
-// ─── Earnings (Issue #181) ────────────────────────────────────────────────────
-
-export interface EarningPayment {
-  id: string;
-  jobId: string;
-  jobTitle: string;
-  amountXlm: string;
-  releasedAt: string;
-  clientAddress: string;
-}
-
-export interface MonthlyEarning {
-  month: string; // "YYYY-MM"
-  totalXlm: number;
-}
-
-export interface EarningsData {
-  totalXlm: string;
-  totalUsdc?: string;
-  payments: EarningPayment[];
-  monthly: MonthlyEarning[];
-}
-
-export async function fetchFreelancerEarnings(
-  publicKey: string,
-): Promise<EarningsData> {
-  const { data } = await api.get<{ success: boolean; data: EarningsData }>(
-    `/api/profiles/${encodeURIComponent(publicKey)}/earnings`,
-  );
-  return data.data;
-}
-
-// ─── Dispute Evidence (Issue #223) ───────────────────────────────────────────
-
-export interface DisputeEvidence {
-  id: string;
-  uploaderAddress: string;
-  fileName: string;
-  fileSize: number;
-  mimeType: string;
-  ipfsCid: string;
-  gatewayUrl: string;
-  createdAt: string;
-}
-
-export interface DisputeDetail {
-  job: {
-    id: string;
-    title: string;
-    status: string;
-    client_address: string;
-    freelancer_address: string;
-    created_at: string;
-  };
-  evidence: DisputeEvidence[];
-}
-
-export async function fetchDisputeDetail(
-  jobId: string,
-): Promise<DisputeDetail> {
-  const { data } = await api.get<{ success: boolean; data: DisputeDetail }>(
-    `/api/disputes/${jobId}`,
-  );
-  return data.data;
-}
-
-export async function uploadDisputeEvidence(
-  jobId: string,
-  file: File,
-  onProgress?: (pct: number) => void,
-): Promise<DisputeEvidence> {
-  const formData = new FormData();
-  formData.append("file", file);
-  const { data } = await api.post<{ success: boolean; data: DisputeEvidence }>(
-    `/api/disputes/${jobId}/evidence`,
-    formData,
-    {
-      headers: { "Content-Type": "multipart/form-data" },
-      timeout: 60000,
-      onUploadProgress: onProgress
-        ? (e) => { if (e.total) onProgress(Math.round((e.loaded / e.total) * 100)); }
-        : undefined,
-    },
-  );
-  return data.data;
-}
-
-export interface SignedEvidenceUrl {
-  url: string;
-  expiresAt: string;
-  fileName: string;
-  mimeType: string;
-}
-
-/** Fetch a 15-minute signed proxy URL for a dispute evidence file (Issue #467). */
-export async function fetchEvidenceSignedUrl(
-  jobId: string,
-  evidenceId: string,
-): Promise<SignedEvidenceUrl> {
-  const { data } = await api.get<{ success: boolean; data: SignedEvidenceUrl }>(
-    `/api/disputes/${jobId}/evidence/${evidenceId}/url`,
-  );
-  return data.data;
-}
-
-export async function verifyPasskeyRegistration(credential: any, name: string) {
-  const { data } = await api.post<{ success: boolean; message: string }>(
-    "/api/webauthn/register-verify",
-    { credential, name },
-  );
-  return data;
-}
-
-export async function fetchPasskeyLoginOptions(publicKey: string) {
-  const { data } = await api.post<{ success: boolean; data: any }>(
-    "/api/webauthn/login-options",
-    { publicKey },
-  );
-  return data.data;
-}
-
-export async function verifyPasskeyLogin(credential: any, publicKey: string) {
-  const { data } = await api.post<{ success: boolean; token: string }>(
-    "/api/webauthn/login-verify",
-    { credential, publicKey },
-  );
-  return data;
-}
-
-export async function fetchPasskeyCredentials(): Promise<PasskeyCredential[]> {
-  const { data } = await api.get<{
-    success: boolean;
-    data: PasskeyCredential[];
-  }>("/api/webauthn/credentials");
-  return data.data;
-}
-
-export async function deletePasskeyCredential(id: string) {
-  await api.delete(`/api/webauthn/credentials/${id}`);
-}
-
-// ─── Developer API ────────────────────────────────────────────────────────────
-
-export interface DeveloperApiKey {
-  id: string;
-  label: string;
-  key_prefix: string;
-  created_at: string;
-  last_used_at: string | null;
-  revoked_at: string | null;
-  rotating_at: string | null;
-  requests_today: number;
-}
-
-export interface CreatedDeveloperApiKey {
-  id: string;
-  label: string;
-  keyPrefix: string;
-  createdAt: string;
-  apiKey: string;
-}
-
-export interface RotatedDeveloperApiKey {
-  id: string;
-  label: string;
-  createdAt: string;
-  rotatingAt: string;
-  apiKey: string;
-}
-
-function buildApiKeyHeaders(apiKey: string) {
-  return { headers: { "X-API-Key": apiKey } };
-}
-
-export async function fetchDeveloperApiKeys(): Promise<DeveloperApiKey[]> {
-  const { data } = await api.get<{ success: boolean; data: DeveloperApiKey[] }>(
-    "/api/developer/keys",
-  );
-  return data.data;
-}
-
-export async function createDeveloperApiKey(
-  label?: string,
-): Promise<CreatedDeveloperApiKey> {
-  const { data } = await api.post<{ success: boolean; data: CreatedDeveloperApiKey }>(
-    "/api/developer/keys",
-    { label },
-  );
-  return data.data;
-}
-
-export async function revokeDeveloperApiKey(id: string): Promise<void> {
-  await api.delete(`/api/developer/keys/${id}`);
-}
-
-export async function rotateDeveloperApiKey(id: string): Promise<RotatedDeveloperApiKey> {
-  const { data } = await api.post<{ success: boolean; data: RotatedDeveloperApiKey }>(
-    `/api/developer/keys/${id}/rotate`,
-  );
-  return data.data;
-}
-
-export async function fetchPublicJobs(apiKey: string, limit = 20) {
-  const { data } = await api.get<{ success: boolean; data: any[] }>(
-    "/api/public/jobs",
-    {
-      params: { limit },
-      ...buildApiKeyHeaders(apiKey),
-    },
-  );
-  return data.data;
-}
-
-export async function fetchPublicJob(apiKey: string, id: string) {
-  const { data } = await api.get<{ success: boolean; data: any }>(
-    `/api/public/jobs/${encodeURIComponent(id)}`,
-    buildApiKeyHeaders(apiKey),
-  );
-  return data.data;
-}
-
-export async function fetchPublicFreelancerProfile(
-  apiKey: string,
-  publicKey: string,
-) {
-  const { data } = await api.get<{ success: boolean; data: any }>(
-    `/api/public/freelancers/${encodeURIComponent(publicKey)}`,
-    buildApiKeyHeaders(apiKey),
-  );
-  return data.data;
-}
-
-// ─── Skill Certificates ─────────────────────────────────────────
-
-export interface CertificateData {
-  id: string;
-  publicKey: string;
-  displayName: string | null;
-  skill: string;
-  score: number;
-  certificateHash: string;
-  ipfsCid: string | null;
-  txHash: string | null;
-  issuedAt: string;
-  verifyUrl: string;
-}
-
-export async function fetchCertificate(id: string): Promise<CertificateData> {
-  const { data } = await api.get<{ success: boolean; data: CertificateData }>(
-    `/api/certificates/${id}`,
-  );
-  return data.data;
-}
-
-export async function fetchUserCertificates(
-  publicKey: string,
-): Promise<CertificateData[]> {
-  const { data } = await api.get<{
-    success: boolean;
-    data: CertificateData[];
-  }>(`/api/certificates/user/${encodeURIComponent(publicKey)}`);
-  return data.data;
-}
-
-// ─── Skill Endorsements ─────────────────────────────────────────
-
-export interface SkillEndorsementData {
-  skill: string;
-  count: number;
-  endorsers: string[];
-}
-
-export async function fetchSkillEndorsements(
-  publicKey: string,
-): Promise<SkillEndorsementData[]> {
-  const { data } = await api.get<{
-    success: boolean;
-    data: SkillEndorsementData[];
-  }>(`/api/profiles/${encodeURIComponent(publicKey)}/endorsements`);
-  return data.data;
-}
-
-export async function endorseSkill(
-  publicKey: string,
-  skill: string,
-): Promise<void> {
-  await api.post(`/api/profiles/${encodeURIComponent(publicKey)}/endorse`, {
-    skill,
-  });
-}
-
-export async function fetchSkillBadges(
-  publicKey: string,
-): Promise<SkillBadge[]> {
-  const { data } = await api.get<{
-    success: boolean;
-    data: SkillBadge[];
-  }>(`/api/assessments/results/${encodeURIComponent(publicKey)}`);
-  return data.data;
-}
-
-// ─── Admin Functions ──────────────────────────────────────────────────────────
-
-export async function fetchAdminMetrics(period: "7d" | "30d" | "90d" = "30d") {
-  const { data } = await api.get<{
-    success: boolean;
-    data: {
-      period: string;
-      platformHealth: {
-        total_jobs: number;
-        open_jobs: number;
-        completed_jobs: number;
-        disputed_jobs: number;
-        completion_rate: number;
-        dispute_rate: number;
-      };
-      userGrowth: {
-        total_users: number;
-        freelancers: number;
-        clients: number;
-        new_users_period: number;
-      };
-      weeklyGrowth: Array<{ week: string; new_users: number }>;
-      financialMetrics: {
-        total_xlm_escrow: number;
-        total_xlm_released: number;
-        avg_job_budget: number;
-        active_escrows: number;
-      };
-      qualityMetrics: {
-        avg_rating: number;
-        total_ratings: number;
-        repeat_hires: number;
-      };
-      disputeMetrics: Array<{
-        week: string;
-        disputes_opened: number;
-        disputes_resolved: number;
-      }>;
-      topEarners: Array<{
-        public_key: string;
-        display_name: string;
-        total_earned_xlm: number;
-        completed_jobs: number;
-        rating: number;
-      }>;
-      jobVolume: Array<{
-        date: string;
-        jobs_created: number;
-        jobs_completed: number;
-      }>;
-    };
-  }>("/api/admin/metrics", { params: { period } });
-  return data.data;
-}
-
-export async function fetchAdminJobReports() {
-  const { data } = await api.get<{ success: boolean; data: any[] }>(
-    "/api/admin/reports/jobs",
-  );
-  return data.data;
-}
-
-export async function fetchAdminDisputes() {
-  const { data } = await api.get<{ success: boolean; data: any[] }>(
-    "/api/admin/disputes",
-  );
-  return data.data;
-}
-
-export async function fetchAdminLogs() {
-  const { data } = await api.get<{ success: boolean; data: any[] }>(
-    "/api/admin/logs",
-  );
-  return data.data;
-}
-
-export async function fetchAuditLogs(params?: {
-  action?: string;
-  resource_type?: string;
-  from?: string;
-  to?: string;
-  limit?: number;
-  after?: string;
-}) {
-  const { data } = await api.get<{
-    success: boolean;
-    data: AuditLogEntry[];
-    nextCursor: string | null;
-  }>("/api/audit", {
-    params,
-  });
-  return { logs: data.data, nextCursor: data.nextCursor };
-}
-
-export async function fetchFrozenWallets() {
-  const { data } = await api.get<{ success: boolean; data: any[] }>(
-    "/api/admin/wallets/frozen",
-  );
-  return data.data;
-}
-
-export async function adminCancelJob(jobId: string, reason: string) {
-  const { data } = await api.patch<{ success: boolean; message: string }>(
-    `/api/admin/jobs/${jobId}/cancel`,
-    { reason },
-  );
-  return data;
-}
-
-export async function freezeWallet(address: string, reason: string) {
-  const { data } = await api.post<{ success: boolean; message: string }>(
-    `/api/admin/wallets/${address}/freeze`,
-    { reason },
-  );
-  return data;
-}
-
-export async function unfreezeWallet(address: string) {
-  const { data } = await api.delete<{ success: boolean; message: string }>(
-    `/api/admin/wallets/${address}/freeze`,
-  );
-  return data;
-}
-
-// ─── Admin Cost Report & Time-Series (Issues #569, #561) ──────────────────────
-
-export async function fetchCostReport() {
-  const { data } = await api.get<{ success: boolean; data: any }>(
-    "/api/admin/cost-report",
-  );
-  return data.data;
-}
-
-export async function generateCostReport() {
-  const { data } = await api.post<{ success: boolean; message: string }>(
-    "/api/admin/cost-report/generate",
-  );
-  return data;
-}
-
-export interface TimeSeriesMetric {
-  metric_name: string;
-  value: number;
-  granularity: string;
-  bucket: string;
-}
-
-export async function fetchTimeSeriesMetrics(params: {
-  metric: string;
-  from?: string;
-  to?: string;
-  granularity?: string;
-}): Promise<TimeSeriesMetric[]> {
-  const { data } = await api.get<{ success: boolean; data: TimeSeriesMetric[] }>(
-    "/api/admin/metrics/time-series",
-    { params },
-  );
-  return data.data;
-}
-
-// ─── Admin API Key Usage Stats (Issue #452) ───────────────────────────────────
-
-export interface ApiKeyUsageEndpoint {
-  endpoint: string;
-  requests: number;
-  lastMinute: string;
-}
-
-export interface ApiKeyUsageRow {
-  id: number;
-  label: string;
-  key_prefix: string;
-  requests_today: number;
-  requests_last_hour: number;
-  endpoint_breakdown: ApiKeyUsageEndpoint[];
-}
-
-export interface ApiKeyUsageStats {
-  lookbackDays: number;
-  keys: ApiKeyUsageRow[];
-}
-
-/**
- * Fetch per-API-key usage statistics for the admin dashboard (Issue #452).
- * Each row includes today's request count, the rolling 60-minute request
- * count, and a per-endpoint breakdown for the most recent activity.
- */
-export async function fetchAdminApiKeyUsage(
-  days = 7,
-): Promise<ApiKeyUsageStats> {
-  const { data } = await api.get<{ success: boolean; data: ApiKeyUsageStats }>(
-    "/api/admin/api-keys/usage",
-    { params: { days } },
-  );
-  return data.data;
-}
-
-// ─── Referrals ────────────────────────────────────────────────────────────────
-
-/**
- * Fetch referral stats and history for a referrer.
- */
-export async function fetchReferralStats(
-  publicKey: string,
-): Promise<ReferralStats> {
-  const { data } = await api.get<{ success: boolean; data: ReferralStats }>(
-    `/api/referrals/${encodeURIComponent(publicKey)}`,
-  );
-  return data.data;
-}
-
-/**
- * Register a referral relationship when a new user signs up via a referral link.
- */
-export async function registerReferral(
-  referrerAddress: string,
-  refereeAddress: string,
-): Promise<void> {
-  await api.post("/api/referrals/register", {
-    referrerAddress,
-    refereeAddress,
-  });
-}
-
-// ─── Saved Searches (Issue #284) ─────────────────────────────────────────────
-
-export interface SavedSearch {
-  id: string;
-  user_address: string;
-  query_params: Record<string, string>;
-  notify_in_app: boolean;
-  notify_email: boolean;
-  last_notified_at: string | null;
-  created_at: string;
-  updated_at: string;
-}
-
-export async function fetchSavedSearches(): Promise<SavedSearch[]> {
-  const { data } = await api.get<{ success: boolean; data: SavedSearch[] }>(
-    "/api/saved-searches"
-  );
-  return data.data;
-}
-
-export async function createSavedSearch(payload: {
-  query_params: Record<string, string>;
-  notify_in_app?: boolean;
-  notify_email?: boolean;
-}): Promise<SavedSearch> {
-  const { data } = await api.post<{ success: boolean; data: SavedSearch }>(
-    "/api/saved-searches",
-    payload
-  );
-  return data.data;
-}
-
-export async function updateSavedSearch(
-  id: string,
-  payload: { notify_in_app?: boolean; notify_email?: boolean }
-): Promise<SavedSearch> {
-  const { data } = await api.patch<{ success: boolean; data: SavedSearch }>(
-    `/api/saved-searches/${id}`,
-    payload
-  );
-  return data.data;
-}
-
-export async function deleteSavedSearch(id: string): Promise<void> {
-  await api.delete(`/api/saved-searches/${id}`);
-}
-
-// ─── Job Boost (Issue #344) ───────────────────────────────────────────────────
-
-/**
- * Notify the backend that a boost payment was made on-chain.
- * The backend sets boosted=true and calculates the expiry from amountXlm.
- */
-export async function boostJob(
-  jobId: string,
-  txHash: string,
-  amountXlm: number,
-): Promise<Job> {
-  const { data } = await api.patch<{ success: boolean; data: Job }>(
-    `/api/jobs/${jobId}/boost`,
-    { txHash, amountXlm },
-  );
-  return data.data;
-}
-
-// ─── Job Invitations (Issue #342) ────────────────────────────────────────────
-
-export interface JobInvitation {
-  id: string;
-  jobId: string;
-  jobTitle: string;
-  jobBudget: string;
-  jobCurrency: string;
-  clientAddress: string;
-  clientName?: string;
-  freelancerAddress: string;
-  status: "pending" | "accepted" | "declined";
-  createdAt: string;
-}
-
-/**
- * Fetch all pending invitations for the authenticated freelancer.
- */
-export async function fetchMyInvitations(): Promise<JobInvitation[]> {
-  const { data } = await api.get<{ success: boolean; data: JobInvitation[] }>(
-    "/api/invitations",
-  );
-  return data.data;
-}
-
-/**
- * Decline a job invitation.
- */
-export async function declineInvitation(invitationId: string): Promise<void> {
-  await api.patch(`/api/invitations/${invitationId}/decline`);
-}
-
-// ─── DAO Governance (#278) ───────────────────────────────────────────────────
-
-export interface DaoProposal {
-  id: string;
-  title: string;
-  description: string;
-  type: "treasury" | "platform" | "parameter" | "arbitration";
-  proposer: string;
-  amount?: string;
-  recipient?: string;
-  votesFor: number;
-  votesAgainst: number;
-  status: "active" | "passed" | "rejected" | "executed";
-  createdAt: string;
-  votingEndsAt: string;
-  quorumPercent?: number;
-  quorumReached?: boolean;
-}
-
-export interface DaoArbitrator {
-  publicKey: string;
-  displayName?: string | null;
-  bio?: string | null;
-  votesReceived: number;
-  disputesResolved: number;
-  electedAt?: string | null;
-}
-
-export async function fetchDaoProposals(status?: string): Promise<DaoProposal[]> {
-  const { data } = await api.get<{ success: boolean; data: DaoProposal[] }>(
-    "/api/dao/proposals",
-    { params: status ? { status } : {} },
-  );
-  return data.data;
-}
-
-export async function createDaoProposal(body: {
-  title: string;
-  description: string;
-  type: DaoProposal["type"];
-  amount?: string;
-  recipient?: string;
-  votingDays?: number;
-}): Promise<DaoProposal> {
-  const { data } = await api.post<{ success: boolean; data: DaoProposal }>(
-    "/api/dao/proposals",
-    body,
-  );
-  return data.data;
-}
-
-export async function voteDaoProposal(
-  proposalId: string,
-  support: boolean,
-  weight: number,
-  txHash?: string,
-): Promise<DaoProposal> {
-  const { data } = await api.post<{ success: boolean; data: DaoProposal }>(
-    `/api/dao/proposals/${proposalId}/vote`,
-    { support, weight, txHash },
-  );
-  return data.data;
-}
-
-export async function fetchDaoTreasury(): Promise<{
-  allocatedXlm: string;
-  activeProposals: number;
-  quorumPercent: number;
-}> {
-  const { data } = await api.get<{
-    success: boolean;
-    data: { allocatedXlm: string; activeProposals: number; quorumPercent: number };
-  }>("/api/dao/treasury");
-  return data.data;
-}
-
-export async function fetchDaoArbitrators(): Promise<{
-  arbitrators: DaoArbitrator[];
-  disputePanel: DaoArbitrator[];
-}> {
-  const { data } = await api.get<{
-    success: boolean;
-    data: { arbitrators: DaoArbitrator[]; disputePanel: DaoArbitrator[] };
-  }>("/api/dao/arbitrators");
-  return data.data;
-}
-
-export async function registerDaoArbitrator(body: {
-  displayName?: string;
-  bio?: string;
-}): Promise<DaoArbitrator> {
-  const { data } = await api.post<{ success: boolean; data: DaoArbitrator }>(
-    "/api/dao/arbitrators",
-    body,
-  );
-  return data.data;
-}
-
-export async function voteDaoArbitrator(
-  arbitratorKey: string,
-  weight: number,
-): Promise<DaoArbitrator[]> {
-  const { data } = await api.post<{ success: boolean; data: DaoArbitrator[] }>(
-    `/api/dao/arbitrators/${arbitratorKey}/vote`,
-    { weight },
-  );
-  return data.data;
-}
-
-/**
- * Accept a job invitation (auto-creates an application).
- */
-export async function acceptInvitation(
-  invitationId: string,
-  proposal: string,
-  bidAmount: string,
-): Promise<Application> {
-  const { data } = await api.post<{ success: boolean; data: Application }>(
-    `/api/invitations/${invitationId}/accept`,
-    { proposal, bidAmount },
-  );
-  return data.data;
-}
-
-// ── Health / Status (#501) ───────────────────────────────────────────────────
-
-export interface HealthStatus {
-  status: "healthy" | "degraded";
-  database: { status: string; latency_ms?: number; message?: string };
-  stellar: { status: string; network?: string; ledger?: number; message?: string };
-  ipfs: { status: string; message?: string };
-  uptime_seconds: number;
-  version: string;
-}
-
-export async function fetchHealthStatus(): Promise<HealthStatus> {
-  const { data } = await api.get<HealthStatus>("/health");
-  return data;
-}
-
-export async function fetchHealthHistory(): Promise<
-  Record<string, { status: string; checkedAt: string }[]>
-> {
-  const { data } = await api.get<{
-    success: boolean;
-    data: Record<string, { status: string; checkedAt: string }[]>;
-  }>("/health/history");
-  return data.data;
-}
-
-export async function subscribeStatusAlerts(email: string): Promise<void> {
-  await api.post("/health/subscribe", { email });
-}
-
-// ── Encryption key + file attachment (#498) ──────────────────────────────────
-
-export async function fetchRecipientEncryptionKey(
-  publicKey: string,
-): Promise<string | null> {
-  const { data } = await api.get<{
-    success: boolean;
-    data: { encryptionPublicKey: string | null };
-  }>(`/api/profiles/${encodeURIComponent(publicKey)}/encryption-key`);
-  return data.data.encryptionPublicKey;
-}
-
-export async function publishMyEncryptionKey(
-  userPublicKey: string,
-  naclPublicKey: string,
-): Promise<void> {
-  await api.put(`/api/profiles/${encodeURIComponent(userPublicKey)}/encryption-key`, {
-    encryptionPublicKey: naclPublicKey,
-  });
-}
-
-export async function uploadMessageAttachment(
-  jobId: string,
-  encryptedBlob: Blob,
-  fileName: string,
-  senderNaclPub: string,
-): Promise<Message> {
-  const formData = new FormData();
-  formData.append("file", encryptedBlob, fileName);
-  formData.append("senderNaclPub", senderNaclPub);
-  const { data } = await api.post<{ success: boolean; data: Message }>(
-    `/api/messages/job/${jobId}/attachments`,
-    formData,
-    { headers: { "Content-Type": "multipart/form-data" }, timeout: 60_000 },
-  );
-  return data.data;
-}
-
-// u2500u2500u2500 Dispute Evidence On-Chain Audit (Issue #448) u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500u2500
-/**
- * Fetch the IPFS CIDs of dispute evidence anchored on-chain for a job.
- * Backed by GET /api/disputes/:jobId/onchain-cids. Returns an empty array
- * if the contract has no entries yet or the network is unreachable.
- */
-export async function fetchDisputeOnchainCids(jobId: string): Promise<string[]> {
-  try {
-    const { data } = await api.get<{ success: boolean; data: { cids: string[] } }>(
-      `/api/disputes/${encodeURIComponent(jobId)}/onchain-cids`,
-    );
-    return Array.isArray(data?.data?.cids) ? data.data.cids : [];
-  } catch {
-    return [];
-  }
-}
