+++ frontend/lib/api/ratings.ts
@@ -1,0 +1,22 @@
+import { api } from "./client";
+import type { Rating } from "@/utils/types";
+
+export async function submitRating(payload: {
+  jobId: string;
+  ratedAddress: string;
+  stars: number;
+  review?: string;
+}) {
+  const { data } = await api.post<{ success: boolean; data: Rating }>(
+    "/api/ratings",
+    payload,
+  );
+  return data.data;
+}
+
+export async function fetchRatings(publicKey: string) {
+  const { data } = await api.get<{ success: boolean; data: Rating[] }>(
+    `/api/ratings/${publicKey}`,
+  );
+  return data.data;
+}
